ENH:adds design to homepage
